chainstore: refactor store.go into more subfiles
